exosphere-runtimes/document-processing-runtime/pyproject.toml
[project]
name = "document-processing-runtime"
version = "0.1.0"
description = "Document processing and question-answering workflow runtime for ExosphereHost"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exospherehost==0.0.7b3",
    "python-dotenv>=1.1.1",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "pydantic>=1.10.0",
    "aiohttp>=3.8.0",
]
